requirements.txt
blinker==1.9.0
click==8.1.8
filelock==3.13.1
Flask==3.1.0
fsspec==2024.2.0
itsdangerous==2.2.0
Jinja2==3.1.3
MarkupSafe==2.1.5
mpmath==1.3.0
networkx==3.2.1
numpy==1.26.3
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
pillow==10.2.0
setuptools==70.0.0
sympy==1.13.1
torch==2.5.1
torchvision==0.20.1
triton==3.1.0
typing_extensions==4.9.0
Werkzeug==3.1.3
